Clean up encapsulation in empty view sizing implementation
