+++ Xamarin.Forms.Platform.iOS/CollectionView/ItemsViewController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using CoreGraphics;
 using Foundation;
 using UIKit;
 
@@ -7,8 +8,6 @@ namespace Xamarin.Forms.Platform.iOS
 	public abstract class ItemsViewController<TItemsView> : UICollectionViewController
 	where TItemsView : ItemsView
 	{
-		public const int HeaderTag = 111;
-		public const int FooterTag = 222;
 		public const int EmptyTag = 333;
 
 		public IItemsViewSource ItemsSource { get; protected set; }
@@ -155,7 +154,7 @@ namespace Xamarin.Forms.Platform.iOS
 			}
 			else
 			{
-				ResizeEmptyView();
+				LayoutEmptyView();
 			}
 		}
 
@@ -292,27 +291,21 @@ namespace Xamarin.Forms.Platform.iOS
 			UpdateEmptyViewVisibility(ItemsSource?.ItemCount == 0);
 		}
 
-		void ResizeEmptyView()
+		protected virtual CGRect DetermineEmptyViewFrame() 
 		{
-			nfloat headerHeight = 0;
-			var headerView = CollectionView.ViewWithTag(HeaderTag);
+			return new CGRect(CollectionView.Frame.X, CollectionView.Frame.Y,
+					CollectionView.Frame.Width, CollectionView.Frame.Height);
+		}
 
-			if (headerView != null)
-				headerHeight = headerView.Frame.Height;
-
-			nfloat footerHeight = 0;
-			var footerView = CollectionView.ViewWithTag(FooterTag);
-
-			if (footerView != null)
-				footerHeight = footerView.Frame.Height;
-
-			var emptyViewFrame = new CoreGraphics.CGRect(CollectionView.Frame.X, CollectionView.Frame.Y, CollectionView.Frame.Width, Math.Abs(CollectionView.Frame.Height - (headerHeight + footerHeight)));
+		void LayoutEmptyView()
+		{
+			var frame = DetermineEmptyViewFrame();	
 
 			if (_emptyUIView != null)
-				_emptyUIView.Frame = emptyViewFrame;
+				_emptyUIView.Frame = frame;
 
-			if (_emptyViewFormsElement != null)
-				_emptyViewFormsElement.Layout(emptyViewFrame.ToRectangle());
+			if (_emptyViewFormsElement != null && ItemsView.LogicalChildren.Contains(_emptyViewFormsElement))
+				_emptyViewFormsElement.Layout(frame.ToRectangle());
 		}
 
 		protected void RemeasureLayout(VisualElement formsElement)
@@ -377,7 +370,7 @@ namespace Xamarin.Forms.Platform.iOS
 
 				_emptyUIView.Tag = EmptyTag;
 				CollectionView.AddSubview(_emptyUIView);
-				ResizeEmptyView();
+				LayoutEmptyView();
 
 				if (_emptyViewFormsElement != null)
 				{
